.github/scripts/requirements.txt
pygithub==2.2.0
ruamel.yaml<0.18.0
ruamel.yaml.string
pyyaml
